added serial version, and callback hook
